streamlined... no array now
I had originally made an array when I was imagining I would need to use loops with coefficients to calculate the ratio.  Turns out it was simpler than I thought, but I kept the array anyway.  Now it is gone and there are just three simple variables for carbs, protein, and fat.